release v3.0.1_BETA by Newui

+++ ui/src/api/api.book.ts
@@ -1,14 +1,21 @@
+import { MOD } from '@/stores/flag'
 import $http from './index'
-import type { Book, CreateBookDto } from '@/types/model/book';
+import local from './local'
 
-const prefix = '/admin/book';
+import type { Book, CreateBookDto } from '@/model/book'
+
+const prefix = '/admin/book'
 
 /**
  * 查询用户信息
  * @returns book
  */
 export function getBook(name: string | undefined): Promise<Book[]> {
-  return $http({ url: prefix + "?name=" + name, method: "get"})
+  if (MOD.value === 'WEB') {
+    return $http({ url: prefix + '?name=' + name, method: 'get' })
+  } else {
+    return local('queryBooks', { userId: localStorage.getItem('userId'), bookName: name })
+  }
 }
 
 /**
@@ -16,17 +23,23 @@ export function getBook(name: string | undefined): Promise<Book[]> {
  * @returns books
  */
 export function getAllBook(): Promise<Book[]> {
-  return $http({ url: "allBook", method: "get"})
+  if (MOD.value === 'WEB') {
+    return $http({ url: 'allBook', method: 'get' })
+  } else {
+    return local('readBooks')
+  }
 }
 
-
-
 /**
-* 注册账本
-* @returns result
-*/
+ * 注册账本
+ * @returns result
+ */
 export function createBook(createDto: CreateBookDto): Promise<Book> {
-  return $http({ url: prefix, method: "post", data: createDto })
+  if (MOD.value === 'WEB') {
+    return $http({ url: prefix, method: 'post', data: createDto })
+  } else {
+    return local('addBook', createDto)
+  }
 }
 
 /**
@@ -34,25 +47,46 @@ export function createBook(createDto: CreateBookDto): Promise<Book> {
  * @returns result
  */
 export function updateBook(book: Book): Promise<Book> {
-  return $http({ url: prefix + "/" + book.id, method: "put", data: book })
+  if (MOD.value === 'WEB') {
+    return $http({ url: prefix + '/' + book.id, method: 'put', data: book })
+  } else {
+    return local('updateBook', book)
+  }
 }
 /**
  * 删除账本
  * @returns result
  */
 export function deleteBook(id: number): Promise<Book> {
-  return $http({ url: prefix + "/" + id, method: "delete" })
+  if (MOD.value === 'WEB') {
+    return $http({ url: prefix + '/' + id, method: 'delete' })
+  } else {
+    return local('deleteBook', id)
+  }
 }
 
+export function checkBook(bookId: string | null): Promise<Book> {
+  return local('checkBook', bookId)
+}
 
 /**
-* 修改账本密钥
-* @returns result
-*/
+ * 修改账本密钥
+ * @returns result
+ */
 export function changeKey(book: Book): Promise<Book> {
-  return $http({ url: prefix + '/changeKey', method: "post", data: book })
+  return $http({ url: prefix + '/changeKey', method: 'post', data: book })
+  // if (MOD.value === 'WEB') {
+  //   return $http({ url: prefix + '/changeKey', method: 'post', data: book })
+  // } else {
+  //   return local('dailyLine', localStorage.getItem('bookId'), query)
+  // }
 }
 
 export async function openBookApi(): Promise<string> {
-  return $http({ url: prefix + '/openBook', method: "post" })
+  return $http({ url: prefix + '/openBook', method: 'post' })
+  // if (MOD.value === 'WEB') {
+  //   return $http({ url: prefix + '/openBook', method: 'post' })
+  // } else {
+  //   return null
+  // }
 }
